Desktop/newsagg/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
trafilatura==2.0.0
openai>=1.25.0
httpx==0.27.0
python-multipart==0.0.9
jinja2==3.1.3
newsapi-python==0.2.7
